move `resolve_fingerprint_values` to fingerprinting utils module
